Dividing slug-file creation and uploading to Heroku into two different maven goals. This enables ci-builds where the slug-file is created once and deployed to more than one stage for testing and continous deployment stuff.
There is a new maven goal called 'heroku:create-slug' and the existing once 'heroku:deploy' respectively 'heroku:deploy-war'.

The slug-filename is configurable in your pom.xml via
<configuration>
  <slugFileName>myapp-prod.tgz</slugFileName>
</configuration>

The target/heroku folder will not be cleared anymore as this is a task of mvn clean.
This way the slug which exists is used and not a new one every time.
The slugfilename should be complete including a suffix. The default tgz suffix is not set in Tar.java anymore.
If no slugFileName is set slug.tgz will be used.