Switch to Cargo.toml based lint configuration
As of the Cargo included in Rust 1.74, lints can now be configured in
Cargo.toml across whole crates:
https://blog.rust-lang.org/2023/11/16/Rust-1.74.0.html
https://doc.rust-lang.org/stable/cargo/reference/manifest.html#the-lints-section

I've also enabled the `unreachable_pub` for parity with our other repos, eg:
https://github.com/heroku/buildpacks-jvm/blob/9c567cafc15206ae8c0cd2aeb2b4cfd171f9589b/Cargo.toml#L18-L26

I've not enabled the various panic/unwrap related lints used in our other
repos, since the Rust usage in this repo is purely for tests, for which panicing
is expected.

GUS-W-14836332.